replace logx with alogx
Conflicts:

	btif/src/btif_core.c
	btif/src/btif_dm.c
	btif/src/btif_hh.c
	btif/src/btif_sock.c
	gki/ulinux/gki_ulinux.c
	hci/include/bt_hci_bdroid.h
	hci/src/bt_hci_bdroid.c
	hci/src/btsnoop.c
	hci/src/hci_h4.c
	hci/src/userial.c
	vendor/libvendor/src/hardware.c
	vendor/libvendor/src/upio.c

+++ btif/src/btif_av.c
@@ -176,7 +176,7 @@ const char *dump_av_sm_state_name(btif_av_state_t state)
 
 const char *dump_av_sm_event_name(btif_av_sm_event_t event)
 {
-    switch(event)
+    switch((int)event)
     {
         CASE_RETURN_STR(BTA_AV_ENABLE_EVT)
         CASE_RETURN_STR(BTA_AV_REGISTER_EVT)
